simplex algorithm works with rationals
The simplex algorithm used to keep track of numbers using javascript floating point numbers. This leads to a big accumulation of errors!
Now it stores numbers as fractions, so everything is precise.
This also puts the objective function at the bottom of the tableau, and draws a line above it.